feat(typing): pass types to class factory, classes and helpers
Class factory accepts an object as a first type parameter. The object specifies the types of the
`dependencyClass`, `dependencyTemplateClass`, `moduleClass`, `moduleFactoryClass` that should be
used within the plugin. Furthermore, also the type of the constructor options object can be
specified with `constructorOptions` property.

Dependency class and its subclassing helper can be
passed an object as a type parameter to specify:
 - `dependencyOptions` - the options object passed
to the Dependency constructor.

ModuleFactory class and its subclassing helper can be passed an
object as a type parameter to specify:
 - `dependency` - the Dependency instance type that is
passed to ModuleFactory.create method
 - `module` - the Module instance type that should be created
and returned by  ModuleFactory.create method.

Module class and its subclassing helper can be
passed an object as a type parameter to specify:
 - `dependency` - the Dependency instance type
that is passed to Module constructor.

Class factory can be passed an object as a type parameter
to specify:
 - `dependencyClass` - type of the `dependencyClass` option passed to the factory
function.
 - `dependencyTemplateClass` - type of the `dependencyTemplateClass` option passed to the
factory function.
 - `moduleClass` - type of the `moduleClass` option passed to the factory
function.
 - `moduleFactoryClass` - type of the `moduleFactoryClass` option passed to the factory
function.
 - `constructorOptions` - type of the options object passed to the instance
constructor.

Additionally added following types to `types.util`:
 - `RequiredKeys` - Type that
accepts and object type and returns a union of keys of required properties or never if there are
none.